gonna scrap the current uniprocessing algorithm that I have and start over. Here goes nothing

+++ myrand.h
@@ -2,7 +2,7 @@
 #define _OTHER_H_
 
 void randInit();
-int getRand();
+float getRand();
 int closeRand();
 
 #endif
